Dice: Add images (#217)
* Update dice.py

Whenever you're asking JARVIS to roll a die, it's just replying with unicode character of dice in python. The problem here the character looks so small on larger resolution desktops or laptops or mobiles.

* Update dice.py

This update includes
1. The images from the following website.  "http://fa2png.io/"
Attachment Template is used this time. 
Thanks for the feedback. I'm new to this Github, would love to learn more.