:speech_balloon: Remove "sir" from the error response (#253)
Close #215 

Edited the file to remove "sir" from the response if the bot is unable to
respond to the query. This still leaves room for someone to add a replace
command to get the sender_name variable, and allow the bot to address the
user by name (such as within the other modules), but for right now this seems
to be the best way to stay gender neutral.
I was looking at using a replace function, but so far have had no such luck.